Fixed:
- UserDTO
- Restaurant-Admin relationship
- Proper Authorization to create, delete, update Restaurants and Dishes

+++ src/Restaurants.Application/Handlers/CommandHandlers/DeleteRestaurantCommandHandler.cs
@@ -3,6 +3,8 @@ using System.Reflection.Metadata.Ecma335;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Commands;
+using Restaurants.Application.Services;
+using Restaurants.Application.Users;
 using Restaurants.Domain.Entitites;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -11,7 +13,8 @@ namespace Restaurants.Application.Handlers;
 
 public class DeleteRestaurantCommandHandler(
     ILogger<DeleteRestaurantCommandHandler> logger,
-    IRestaurantsRepository repository)
+    IRestaurantsRepository repository,
+    IUserContext userContext)
     : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
@@ -24,6 +27,10 @@ public class DeleteRestaurantCommandHandler(
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
         }
 
+        var user = userContext.GetCurrentUser();
+        if(!user.IsAuthorized(restaurant)){
+            throw new ForbidException(user.Id, nameof(this.GetType), nameof(Restaurant), request.Id.ToString());
+        }
         await repository.DeleteRestaurantAsync(restaurant);
     }
 }
